Able to delete users - passing all the tests for delete

+++ server/resources/events/events-handler.ts
@@ -67,12 +67,21 @@ export const updateEvent = async (req: Request, res: Response) => {
 
 export const deleteEvent = async (req: Request, res: Response) => {
   try {
-    const event = await EventModel.findByIdAndDelete(req.params.id);
+    // const event = await EventModel.findByIdAndDelete(req.params.id);
+    const event = await EventModel.findOne({ _id: req.params.id });
     if (!event) {
       res.status(404).json('Event not found');
       return;
     }
 
+    if (
+      event.author.toString() !== req.session!.user._id.toString() &&
+      !req.session!.user.isAdmin
+    ) {
+      res.status(403).json('Not authorized to delete this event');
+      return;
+    }
+    await EventModel.findByIdAndDelete(req.params.id);
     res.status(204).json('Event deleted');
   } catch (error) {
     res.status(500).json(error);
